# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/posts":{"post":{"summary":"Create Post","description":"Create a post. Username is required. You must provide at least content or an image (or both).\n\nExpects multipart/form-data with:\n- username (form field, required)\n- content  (form field, optional)\n- image    (file field, optional, PNG or JPG only)","operationId":"createPost","requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_createPost"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Posts","operationId":"get_posts_posts_get","parameters":[{"name":"user","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter posts by exact username","title":"User"},"description":"Filter posts by exact username"},{"name":"order_by","in":"query","required":false,"schema":{"type":"string","pattern":"^(created_at|id)$","default":"created_at","title":"Order By"}},{"name":"order_dir","in":"query","required":false,"schema":{"type":"string","pattern":"^(asc|desc)$","default":"desc","title":"Order Dir"}},{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":1000,"minimum":1},{"type":"null"}],"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostOut"},"title":"Response Get Posts Posts Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/posts/search":{"get":{"summary":"Search Posts","operationId":"search_posts_posts_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","title":"Search Query","description":"Search in post content and usernames"},"description":"Search in post content and usernames"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostOut"},"title":"Response Search Posts Posts Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/posts/{post_id}":{"get":{"summary":"Get Post","operationId":"get_post_posts__post_id__get","parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","title":"Post Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/posts/{post_id}/describe":{"post":{"summary":"Describe Post","operationId":"describe_post_posts__post_id__describe_post","parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","title":"Post Id"}}],"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_createPost":{"properties":{"username":{"type":"string","title":"Username"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"image":{"anyOf":[{"type":"string","format":"binary"},{"type":"null"}],"title":"Image"}},"type":"object","required":["username"],"title":"Body_createPost"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PostOut":{"properties":{"id":{"type":"integer","title":"Id"},"image_filename":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Filename"},"image_status":{"type":"string","enum":["PENDING","READY","FAILED"],"title":"Image Status"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"username":{"type":"string","title":"Username"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"image_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Description"},"description_status":{"type":"string","enum":["NONE","PENDING","READY","FAILED"],"title":"Description Status"}},"type":"object","required":["id","image_status","username","created_at","description_status"],"title":"PostOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}